+++ components/src/user/panel.vue
@@ -11,6 +11,7 @@ const { user, leave } = useUser()
   router-link(:to="`/users/${user.pub}`")
     account-avatar
   .text-2xl.mx-2.font-bold {{ user?.name }}
+  account-mate(:pub="user.pub")
   .flex-1 
   util-pulse(:blink="user.blink")
   button.p-2.text-2xl(@click="user.db.get('safe').get('saved').put(!user.safe.saved)")